Cap 2 - Travel Your Way/travel your way/requirements.txt
Flask==2.1.0
Werkzeug==2.0.2
flask-wtf==0.15.1
wtforms==2.3.3
requests==2.26.0
APScheduler==3.8.1
Flask-SQLAlchemy==3.0.1
Flask-Bcrypt==0.7.1
